wiz-mcp/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wiz-mcp-server"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server for the Wiz API"
readme = "README.md"
authors = [
    {name = "Wiz Security", email = "info@wiz.io"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]==1.6.0",
    "httpx==0.28.1",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
]

[project.urls]
"Homepage" = "https://github.com/wiz-sec/wiz-mcp"
"Bug Tracker" = "https://github.com/wiz-sec/wiz-mcp/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
